+++ src/barelymusician/dsp/dsp_utils.cpp
@@ -16,12 +16,12 @@ float AmplitudeFromDecibels(float decibels) {
   return 0.0f;
 }
 
-float BeatsFromSamples(int samples, int num_samples_per_beat) {
+double BeatsFromSamples(int samples, int num_samples_per_beat) {
   DCHECK_GE(num_samples_per_beat, 0);
   return (num_samples_per_beat > 0)
-             ? static_cast<float>(samples) /
-                   static_cast<float>(num_samples_per_beat)
-             : 0.0f;
+             ? static_cast<double>(samples) /
+                   static_cast<double>(num_samples_per_beat)
+             : 0.0;
 }
 
 float DecibelsFromAmplitude(float amplitude) {
@@ -43,9 +43,9 @@ float GetFilterCoefficient(int sample_rate, float cuttoff_frequency) {
   return 0.0f;
 }
 
-int SamplesFromBeats(float beats, int num_samples_per_beat) {
+int SamplesFromBeats(double beats, int num_samples_per_beat) {
   DCHECK_GE(num_samples_per_beat, 0);
-  return static_cast<int>(beats * static_cast<float>(num_samples_per_beat));
+  return static_cast<int>(beats * static_cast<double>(num_samples_per_beat));
 }
 
 }  // namespace barelyapi
