+++ src/barelymusician/dsp/dsp_utils.h
@@ -8,7 +8,7 @@ namespace barelyapi {
 // @param samples Number of samples.
 // @param num_samples_per_beat Number of samples per beat.
 // @return Fractional number of beats.
-float BeatsFromSamples(int samples, int num_samples_per_beat);
+double BeatsFromSamples(int samples, int num_samples_per_beat);
 
 // Converts a value from dB to linear amplitude.
 //
@@ -34,7 +34,7 @@ float GetFilterCoefficient(int sample_rate, float cuttoff_frequency);
 // @param beats Fractional number of beats.
 // @param num_samples_per_beat Number of samples per beat.
 // @return Number of samples.
-int SamplesFromBeats(float beats, int num_samples_per_beat);
+int SamplesFromBeats(double beats, int num_samples_per_beat);
 
 }  // namespace barelyapi
